refactor(core): extract ANE and sidecar model downloads
Move ANE package presence/unpack into download/ane.rs and the
speaker / punctuation / VAD ensure helpers into download/sidecars.rs.
