+++ crates/gigastt-core/src/model/download.rs
@@ -13,17 +13,7 @@ use std::os::fd::AsRawFd;
 
 use super::variant::ModelVariant;
 #[cfg(feature = "net")]
-use super::variant::{
-    PREQUANT_RELEASE_BASE, PUNCT_FILES, PUNCT_HF_REPO, VAD_MODEL_SHA256, VAD_MODEL_URL,
-};
-
-#[cfg(all(feature = "ane", not(feature = "net")))]
-use super::variant::ANE_BUCKETS;
-#[cfg(all(feature = "net", feature = "ane"))]
-use super::variant::{ANE_BUCKETS, ANE_RELEASE_BASE, ANE_TAR_CHECKSUMS};
-
-#[cfg(all(feature = "diarization", feature = "net"))]
-use super::variant::{SPEAKER_HF_REPO, SPEAKER_MODEL_FILE, SPEAKER_MODEL_SHA256};
+use super::variant::PREQUANT_RELEASE_BASE;
 
 /// Stream a file and return its lowercase SHA-256 hex digest.
 ///
@@ -363,331 +353,6 @@ pub async fn ensure_prequantized_model_variant(
     Ok(variant)
 }
 
-/// Directory name of the unpacked `.mlpackage` for a given mel bucket.
-#[cfg(feature = "ane")]
-pub fn ane_package_dir_name(bucket: usize) -> String {
-    format!("gigaam_v3_encoder_{bucket}.mlpackage")
-}
-
-/// Filename of the published `.tar` artifact for a given mel bucket.
-#[cfg(all(feature = "net", feature = "ane"))]
-pub(super) fn ane_tar_name(bucket: usize) -> String {
-    format!("{}.tar", ane_package_dir_name(bucket))
-}
-
-/// Pinned `.tar` SHA-256 for `bucket`, or `None` when unreleased (the empty
-/// sentinel in [`ANE_TAR_CHECKSUMS`]).
-#[cfg(all(feature = "net", feature = "ane"))]
-fn ane_tar_checksum(bucket: usize) -> Option<&'static str> {
-    ANE_TAR_CHECKSUMS
-        .iter()
-        .find(|(b, _)| *b == bucket)
-        .and_then(|(_, sum)| if sum.is_empty() { None } else { Some(*sum) })
-}
-
-/// Return the default ANE-model directory (`~/.gigastt/models/ane/`), a sibling
-/// of [`default_model_dir`].
-///
-/// Holds the per-bucket palettized Core ML encoder packages the macOS ANE
-/// backend runs. The packages auto-download via [`ensure_ane_packages`] when the
-/// ANE path is requested (`gigastt download --ane`).
-#[cfg(feature = "ane")]
-pub fn default_ane_model_dir() -> String {
-    home_dir()
-        .map(|h| {
-            h.join(".gigastt")
-                .join("models")
-                .join("ane")
-                .to_string_lossy()
-                .into_owned()
-        })
-        .unwrap_or_else(|| ".gigastt/models/ane".into())
-}
-
-/// True when `pkg_dir` is a fully-formed Core ML `.mlpackage` directory.
-///
-/// Requires every structurally-load-bearing member, not just the
-/// `Manifest.json` marker: the manifest, the serialized model spec, and the
-/// weights blob. A package missing any of these cannot load on the ANE, so
-/// treating it as "present" would wedge the download path forever. Observed
-/// layout (real published package, all three buckets identical):
-///   `Manifest.json`
-///   `Data/com.apple.CoreML/model.mlmodel`
-///   `Data/com.apple.CoreML/weights/weight.bin`
-#[cfg(feature = "ane")]
-pub fn ane_package_complete(pkg_dir: &Path) -> bool {
-    pkg_dir.is_dir()
-        && pkg_dir.join("Manifest.json").is_file()
-        && pkg_dir
-            .join("Data")
-            .join("com.apple.CoreML")
-            .join("model.mlmodel")
-            .is_file()
-        && pkg_dir
-            .join("Data")
-            .join("com.apple.CoreML")
-            .join("weights")
-            .join("weight.bin")
-            .is_file()
-}
-
-/// True when every bucket's unpacked `.mlpackage` is present and complete in
-/// `dir` (see [`ane_package_complete`] for the structural requirements).
-#[cfg(feature = "ane")]
-pub fn is_ane_present(dir: &Path) -> bool {
-    ANE_BUCKETS
-        .iter()
-        .all(|&b| ane_package_complete(&dir.join(ane_package_dir_name(b))))
-}
-
-/// Ensure the per-bucket ANE Core ML encoder packages exist in `model_dir`,
-/// downloading each bucket's deterministic `.tar` from the pinned GitHub Release
-/// and unpacking it to reconstruct the `.mlpackage` directory.
-///
-/// Each `.tar` is SHA-256-verified against [`ANE_TAR_CHECKSUMS`] (one digest =
-/// content identity = download pin) before being unpacked with the `tar` crate's
-/// default path-traversal guard, then the `.tar` is removed. Buckets whose
-/// `.mlpackage` is already present are skipped. Reuses the same streaming
-/// download + atomic-rename + lock infra as [`ensure_prequantized_model_variant`].
-///
-/// Bails with a clear message when the release is not yet published (sentinel
-/// checksums).
-#[cfg(all(feature = "net", feature = "ane"))]
-pub async fn ensure_ane_packages(model_dir: &str) -> Result<()> {
-    let dir = Path::new(model_dir);
-
-    if is_ane_present(dir) {
-        tracing::info!("ANE encoder packages found at {model_dir}");
-        return Ok(());
-    }
-
-    std::fs::create_dir_all(dir).context("Failed to create ANE model directory")?;
-
-    #[cfg(unix)]
-    let _lock = acquire_download_lock(dir)?;
-
-    // Re-check after acquiring the lock in case another process finished.
-    if is_ane_present(dir) {
-        tracing::info!("ANE encoder packages found at {model_dir} after lock");
-        return Ok(());
-    }
-
-    tracing::info!("Downloading ANE encoder packages from {ANE_RELEASE_BASE}...");
-
-    for &bucket in ANE_BUCKETS {
-        let pkg_name = ane_package_dir_name(bucket);
-        if ane_package_complete(&dir.join(&pkg_name)) {
-            continue;
-        }
-        let checksum = require_ane_tar_checksum(bucket)?;
-        let tar_name = ane_tar_name(bucket);
-        let tar_dest = dir.join(&tar_name);
-        let url = format!("{ANE_RELEASE_BASE}/{tar_name}");
-        stream_to_partial_then_finalize(&url, &tar_dest, Some(checksum), &tar_name).await?;
-
-        // Extract atomically: unpack into a unique staging dir on the SAME
-        // filesystem, then rename the reconstructed package into place so the
-        // present-check only ever observes a fully-formed `.mlpackage`. A torn
-        // unpack (disk-full / SIGKILL) leaves only the staging dir + the `.tar`,
-        // both of which we remove on every error path so a retry starts clean.
-        tracing::info!("Unpacking {tar_name} into {model_dir}");
-        if let Err(e) = extract_ane_tar_atomic(&tar_dest, dir, &pkg_name) {
-            let _ = std::fs::remove_file(&tar_dest);
-            return Err(e);
-        }
-        // `.tar` removed only AFTER a successful rename, so a failed run above
-        // retains it for retry.
-        std::fs::remove_file(&tar_dest)
-            .with_context(|| format!("Failed to remove {}", tar_dest.display()))?;
-    }
-
-    tracing::info!("ANE encoder packages download complete");
-    Ok(())
-}
-
-/// Unpack `tar_dest` into a unique staging dir under `dir` (same filesystem →
-/// atomic rename), then move the reconstructed `<pkg_name>` package into
-/// `dir/<pkg_name>` with a single `rename`. The package only ever appears at
-/// its final path fully-formed.
-///
-/// The deterministic `.tar`'s arcnames are prefixed with `<pkg_name>/`, so the
-/// reconstructed package lands at `staging/<pkg_name>`. `tar::Archive::unpack`
-/// keeps its default path-traversal guard (entries escaping the target are
-/// rejected). On any failure the staging dir is removed before bailing so a
-/// retry starts clean; the caller removes the `.tar`.
-#[cfg(all(feature = "net", feature = "ane"))]
-pub(super) fn extract_ane_tar_atomic(tar_dest: &Path, dir: &Path, pkg_name: &str) -> Result<()> {
-    // Unique per-process staging dir, same pid+nanos scheme as
-    // `partial_path_unique`, kept under `dir` so the final rename is atomic.
-    let stamp = std::time::SystemTime::now()
-        .duration_since(std::time::UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_nanos();
-    let staging = dir.join(format!(".extract.{}.{}", std::process::id(), stamp));
-
-    // Best-effort staging cleanup before any `?`-bail.
-    let cleanup_staging = || {
-        let _ = std::fs::remove_dir_all(&staging);
-    };
-
-    if let Err(e) = std::fs::create_dir_all(&staging)
-        .with_context(|| format!("Failed to create staging dir {}", staging.display()))
-    {
-        cleanup_staging();
-        return Err(e);
-    }
-
-    let unpack = (|| -> Result<()> {
-        let tar_file = std::fs::File::open(tar_dest)
-            .with_context(|| format!("Failed to open {}", tar_dest.display()))?;
-        tar::Archive::new(tar_file)
-            .unpack(&staging)
-            .with_context(|| format!("Failed to unpack {}", tar_dest.display()))?;
-
-        let src = staging.join(pkg_name);
-        let dest = dir.join(pkg_name);
-        // A torn package from a prior aborted run (the strengthened present-check
-        // now rejects it) must be cleared before rename, or `rename` fails with
-        // "directory not empty".
-        if dest.exists() {
-            std::fs::remove_dir_all(&dest)
-                .with_context(|| format!("Failed to remove stale {}", dest.display()))?;
-        }
-        std::fs::rename(&src, &dest)
-            .with_context(|| format!("Failed to rename {} -> {}", src.display(), dest.display()))?;
-        Ok(())
-    })();
-
-    cleanup_staging();
-    unpack
-}
-
-/// Resolve the pinned `.tar` checksum for `bucket`, bailing with the
-/// not-yet-published message when it is a sentinel. Factored out so the
-/// sentinel-bail branch is unit-testable without the network / async path.
-#[cfg(all(feature = "net", feature = "ane"))]
-pub(super) fn require_ane_tar_checksum(bucket: usize) -> Result<&'static str> {
-    ane_tar_checksum(bucket).ok_or_else(|| {
-        anyhow::anyhow!(
-            "ANE encoder release not yet published; run the Release ANE workflow \
-             (release-ane.yml), then pin the per-bucket .tar SHA-256 from \
-             SHA256SUMS.txt in ANE_TAR_CHECKSUMS"
-        )
-    })
-}
-
-/// Ensure the speaker diarization model exists in `model_dir`, downloading from HuggingFace if missing.
-///
-/// Downloads `wespeaker_resnet34.onnx` from `onnx-community/wespeaker-voxceleb-resnet34-LM`
-/// into `<model_dir>/wespeaker_resnet34.onnx.partial`, verifies its SHA-256 against
-/// `SPEAKER_MODEL_SHA256`, and atomically renames it into place. On checksum mismatch or
-/// crash the final path is never observable, so a subsequent `ensure_speaker_model` call
-/// will re-download from scratch rather than loading a tampered model.
-#[cfg(feature = "diarization")]
-#[cfg(feature = "net")]
-pub async fn ensure_speaker_model(model_dir: &str) -> Result<()> {
-    let dir = Path::new(model_dir);
-    let final_dest = dir.join(SPEAKER_MODEL_FILE);
-
-    if final_dest.exists() {
-        tracing::info!("Speaker model found at {}", final_dest.display());
-        return Ok(());
-    }
-
-    tracing::info!("Speaker model not found, downloading from HuggingFace...");
-    std::fs::create_dir_all(dir).context("Failed to create model directory")?;
-
-    let url = format!("https://huggingface.co/{SPEAKER_HF_REPO}/resolve/main/onnx/model.onnx");
-    stream_to_partial_then_finalize(
-        &url,
-        &final_dest,
-        Some(SPEAKER_MODEL_SHA256),
-        SPEAKER_MODEL_FILE,
-    )
-    .await
-}
-
-/// Ensure the optional punctuation model exists in `punct_model_dir`,
-/// downloading any missing files from the `ekhodzitsky/rupunct-small-onnx`
-/// HuggingFace repo (public, MIT).
-///
-/// Downloads the three files the punctuation pass needs
-/// (`rupunct_small_int8.onnx`, `tokenizer.json`, `config.json`) — only those
-/// not already present — using the same streaming-download + atomic-rename +
-/// SHA-256 infra as the main model download. Files already on disk are left
-/// untouched, so a second call is a no-op (no re-download).
-///
-/// The pass is strictly optional: callers treat a download error as
-/// "punctuation unavailable" and proceed with bare text.
-#[cfg(feature = "net")]
-pub async fn ensure_punct_model(punct_model_dir: &str) -> Result<()> {
-    let dir = Path::new(punct_model_dir);
-
-    if PUNCT_FILES.iter().all(|(file, _)| dir.join(file).exists()) {
-        tracing::info!("Punctuation model found at {punct_model_dir}");
-        return Ok(());
-    }
-
-    tracing::info!("Punctuation model not found, downloading from HuggingFace...");
-    std::fs::create_dir_all(dir).context("Failed to create punctuation model directory")?;
-
-    #[cfg(unix)]
-    let _lock = acquire_download_lock(dir)?;
-
-    for (file, sha256) in PUNCT_FILES {
-        let final_dest = dir.join(file);
-        if final_dest.exists() {
-            continue;
-        }
-        let url = format!("https://huggingface.co/{PUNCT_HF_REPO}/resolve/main/{file}");
-        stream_to_partial_then_finalize(&url, &final_dest, Some(sha256), file).await?;
-    }
-
-    tracing::info!("Punctuation model download complete");
-    Ok(())
-}
-
-/// Ensure the optional Silero VAD model exists in `vad_model_dir`, downloading
-/// it from the pinned Silero release (MIT) if missing.
-///
-/// Uses the same streaming-download + atomic-rename + SHA-256 infra as the main
-/// model download. A file already on disk is left untouched (no re-download).
-///
-/// VAD is strictly optional: callers treat a download error as "VAD
-/// unavailable" and proceed without silence skipping / VAD endpointing.
-#[cfg(feature = "net")]
-pub async fn ensure_vad_model(vad_model_dir: &str) -> Result<()> {
-    let dir = Path::new(vad_model_dir);
-    let final_dest = dir.join(crate::vad::VAD_MODEL_FILE);
-
-    if final_dest.exists() {
-        tracing::info!("VAD model found at {}", final_dest.display());
-        return Ok(());
-    }
-
-    tracing::info!("VAD model not found, downloading from {VAD_MODEL_URL}...");
-    std::fs::create_dir_all(dir).context("Failed to create VAD model directory")?;
-
-    #[cfg(unix)]
-    let _lock = acquire_download_lock(dir)?;
-
-    // Another process may have finished while we waited for the lock.
-    if final_dest.exists() {
-        return Ok(());
-    }
-
-    stream_to_partial_then_finalize(
-        VAD_MODEL_URL,
-        &final_dest,
-        Some(VAD_MODEL_SHA256),
-        crate::vad::VAD_MODEL_FILE,
-    )
-    .await?;
-
-    tracing::info!("VAD model download complete");
-    Ok(())
-}
-
 /// True when every downloaded file for `variant` is present in `dir`.
 ///
 /// Checks the *downloaded* set (FP32 encoder, decoder, joiner, vocab); the
@@ -735,3 +400,17 @@ pub(super) fn partial_path(final_path: &Path) -> std::path::PathBuf {
 pub(super) mod fetch;
 #[cfg(feature = "net")]
 use fetch::{download_file, stream_to_partial_then_finalize};
+
+#[cfg(feature = "ane")]
+pub(super) mod ane;
+#[cfg(all(feature = "net", feature = "ane"))]
+pub use ane::ensure_ane_packages;
+#[cfg(feature = "ane")]
+pub use ane::{ane_package_complete, ane_package_dir_name, default_ane_model_dir, is_ane_present};
+
+#[cfg(feature = "net")]
+mod sidecars;
+#[cfg(all(feature = "net", feature = "diarization"))]
+pub use sidecars::ensure_speaker_model;
+#[cfg(feature = "net")]
+pub use sidecars::{ensure_punct_model, ensure_vad_model};
